refactor(cli): pull abstract handler out of types
