+++ packages/@expo/cli/src/start/server/metro/debugging/messageHandlers/PageReload.ts
@@ -1,7 +1,8 @@
 import type { Protocol } from 'devtools-protocol';
 
 import type { MetroBundlerDevServer } from '../../MetroBundlerDevServer';
-import { type CdpMessage, type Connection, type DebuggerRequest, MessageHandler } from '../types';
+import { MessageHandler } from '../MessageHandler';
+import type { CdpMessage, Connection, DebuggerRequest } from '../types';
 
 export class PageReloadHandler extends MessageHandler {
   private metroBundler: MetroBundlerDevServer;
