+++ packages/@expo/cli/src/start/server/metro/debugging/messageHandlers/VscodeDebuggerGetPossibleBreakpoints.ts
@@ -1,12 +1,8 @@
 import type Protocol from 'devtools-protocol';
 
+import { MessageHandler } from '../MessageHandler';
 import { getDebuggerType } from '../getDebuggerType';
-import {
-  type CdpMessage,
-  type DebuggerRequest,
-  type DeviceResponse,
-  MessageHandler,
-} from '../types';
+import type { CdpMessage, DebuggerRequest, DeviceResponse } from '../types';
 
 /**
  * Hermes doesn't seem to handle this request, but `locations` have to be returned.
