+++ packages/@expo/cli/src/start/server/metro/debugging/messageHandlers/__tests__/VscodeRuntimeGetProperties.test.ts
@@ -1,6 +1,6 @@
 import { mockConnection } from './testUtilts';
 import { getDebuggerType } from '../../getDebuggerType';
-import { type DebuggerRequest } from '../../types';
+import type { DebuggerRequest } from '../../types';
 import {
   type RuntimeGetProperties,
   VscodeRuntimeGetPropertiesHandler,
